+++ pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>io.nsq</groupId>
-  <artifactId>client</artifactId>
+  <groupId>com.github.brainlag.nsq</groupId>
+  <artifactId>nsq-client</artifactId>
   <version>1.0.0-SNAPSHOT</version>
 
   <name>JavaNSQClient</name>
@@ -21,8 +21,6 @@
     <developer>
       <name>Peter Nimmervoll</name>
       <email>nimmervoll.peter@gmail.com</email>
-      <organization>nsq</organization>
-      <organizationUrl>nsq.io</organizationUrl>
     </developer>
   </developers>
   <scm>
@@ -94,6 +92,17 @@
       </build>
     </profile>
   </profiles>
+
+  <distributionManagement>
+    <snapshotRepository>
+      <id>ossrh</id>
+      <url>https://oss.sonatype.org/content/repositories/snapshots</url>
+    </snapshotRepository>
+    <repository>
+      <id>ossrh</id>
+      <url>https://oss.sonatype.org/service/local/staging/deploy/maven2/</url>
+    </repository>
+  </distributionManagement>
 
   <dependencies>
     <dependency>
