+++ src/main/java/com/github/brainlag/nsq/NSQCommand.java
@@ -1,0 +1,47 @@
+package com.github.brainlag.nsq;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class NSQCommand {
+
+	public static NSQCommand instance(String line) {
+		NSQCommand n = new NSQCommand();
+		n.setLine(line);
+		return n;
+	}
+	
+	public static NSQCommand instance(String line, byte[] bytes) {
+		NSQCommand n = instance(line);
+		n.addBytes(bytes);
+		return n;
+	}
+	
+	String line;
+	List<byte[]> data = new ArrayList<>();
+	
+	public void addBytes(byte[] bytes) {
+		data.add(bytes);
+	}
+	
+	public String getLine() {
+		return line;
+	}
+	public void setLine(String line) {
+		if (!line.endsWith("\n")) {
+			line = line +"\n";
+		}
+		
+		this.line = line;
+	}
+	public List<byte[]> getData() {
+		return data;
+	}
+	public void setData(List<byte[]> data) {
+		this.data = data;
+	}
+	
+	public String toString() {
+        return this.getLine().trim();
+    }
+}
