+++ src/main/java/com/github/brainlag/nsq/ServerAddress.java
@@ -1,0 +1,40 @@
+package com.github.brainlag.nsq;
+
+import java.util.Objects;
+
+public class ServerAddress {
+
+    public ServerAddress(final String host, final int port) {
+        this.host = host;
+        this.port = port;
+    }
+
+    public String getHost() {
+        return host;
+    }
+
+    public int getPort() {
+        return port;
+    }
+
+    public String toString() {
+        return host + ":" + port;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        final ServerAddress that = (ServerAddress) o;
+        return Objects.equals(port, that.port) &&
+                Objects.equals(host, that.host);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(host, port);
+    }
+
+    private String host;
+    private int port;
+}
