+++ src/main/java/com/github/brainlag/nsq/callbacks/NSQErrorCallback.java
@@ -1,0 +1,9 @@
+package com.github.brainlag.nsq.callbacks;
+
+import com.github.brainlag.nsq.exceptions.NSQException;
+
+@FunctionalInterface
+public interface NSQErrorCallback {
+
+    void error(NSQException x);
+}
