+++ src/main/java/com/github/brainlag/nsq/callbacks/NSQMessageCallback.java
@@ -1,0 +1,9 @@
+package com.github.brainlag.nsq.callbacks;
+
+import com.github.brainlag.nsq.NSQMessage;
+
+@FunctionalInterface
+public interface NSQMessageCallback {
+
+	public void message(NSQMessage message);
+}
