+++ src/main/java/com/github/brainlag/nsq/exceptions/BadMessageException.java
@@ -1,0 +1,8 @@
+package com.github.brainlag.nsq.exceptions;
+
+public class BadMessageException extends NSQException {
+
+	public BadMessageException(String message) {
+		super(message);
+	}
+}
