+++ src/main/java/com/github/brainlag/nsq/exceptions/BadTopicException.java
@@ -1,0 +1,8 @@
+package com.github.brainlag.nsq.exceptions;
+
+public class BadTopicException extends NSQException {
+
+	public BadTopicException(String message) {
+		super(message);
+	}
+}
