+++ src/main/java/com/github/brainlag/nsq/exceptions/DisconnectedException.java
@@ -1,0 +1,8 @@
+package com.github.brainlag.nsq.exceptions;
+
+public class DisconnectedException extends NSQException {
+
+	public DisconnectedException(String message, Throwable cause) {
+		super(message, cause);
+	}
+}
