+++ src/main/java/com/github/brainlag/nsq/exceptions/NoConnectionsException.java
@@ -1,0 +1,12 @@
+package com.github.brainlag.nsq.exceptions;
+
+public class NoConnectionsException extends NSQException {
+
+	public NoConnectionsException(String message) {
+		super(message);
+	}
+
+	public NoConnectionsException(String message, Throwable cause) {
+		super(message, cause);
+	}
+}
