--- src/main/java/io/nsq/callbacks/NSQErrorCallback.java
@@ -1,9 +1,0 @@
-package io.nsq.callbacks;
-
-import io.nsq.exceptions.NSQException;
-
-@FunctionalInterface
-public interface NSQErrorCallback {
-
-    void error(NSQException x);
-}
