--- src/main/java/io/nsq/callbacks/NSQMessageCallback.java
@@ -1,9 +1,0 @@
-package io.nsq.callbacks;
-
-import io.nsq.NSQMessage;
-
-@FunctionalInterface
-public interface NSQMessageCallback {
-
-	public void message(NSQMessage message);
-}
