--- src/main/java/io/nsq/exceptions/BadMessageException.java
@@ -1,8 +1,0 @@
-package io.nsq.exceptions;
-
-public class BadMessageException extends NSQException {
-
-	public BadMessageException(String message) {
-		super(message);
-	}
-}
