--- src/main/java/io/nsq/exceptions/NSQException.java
@@ -1,25 +1,0 @@
-package io.nsq.exceptions;
-
-import io.nsq.frames.ErrorFrame;
-
-public class NSQException extends Exception {
-
-    public NSQException(String message) {
-        super(message);
-    }
-
-    public NSQException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public static NSQException of(ErrorFrame frame) {
-        String err = frame.getErrorMessage();
-        if (err.startsWith("E_BAD_TOPIC")) {
-            return new BadTopicException(err);
-        }
-        if (err.startsWith("E_BAD_MESSAGE")) {
-            return new BadMessageException(err);
-        }
-        return new NSQException(err);
-    }
-}
