--- src/main/java/io/nsq/exceptions/NoConnectionsException.java
@@ -1,12 +1,0 @@
-package io.nsq.exceptions;
-
-public class NoConnectionsException extends NSQException {
-
-	public NoConnectionsException(String message) {
-		super(message);
-	}
-
-	public NoConnectionsException(String message, Throwable cause) {
-		super(message, cause);
-	}
-}
