--- src/main/java/io/nsq/pool/ConnectionPoolFactory.java
@@ -1,54 +1,0 @@
-package io.nsq.pool;
-
-import io.netty.bootstrap.Bootstrap;
-import io.netty.channel.ChannelFuture;
-import io.netty.channel.nio.NioEventLoopGroup;
-import io.netty.channel.socket.nio.NioSocketChannel;
-import io.nsq.Connection;
-import io.nsq.NSQCommand;
-import io.nsq.NSQConfig;
-import io.nsq.ServerAddress;
-import io.nsq.netty.NSQClientInitializer;
-import org.apache.commons.pool2.BaseKeyedPooledObjectFactory;
-import org.apache.commons.pool2.PooledObject;
-import org.apache.commons.pool2.impl.DefaultPooledObject;
-
-import java.util.concurrent.ExecutorService;
-
-public class ConnectionPoolFactory extends BaseKeyedPooledObjectFactory<ServerAddress, Connection> {
-    private Bootstrap bootstrap;
-    private ExecutorService executor;
-    private NSQConfig config;
-
-
-    public ConnectionPoolFactory(NSQConfig config, ExecutorService executor) {
-        bootstrap = new Bootstrap();
-        bootstrap.group(new NioEventLoopGroup());
-        bootstrap.channel(NioSocketChannel.class);
-        bootstrap.handler(new NSQClientInitializer());
-        this.executor = executor;
-        this.config = config;
-    }
-
-    @Override
-    public Connection create(final ServerAddress serverAddress) throws Exception {
-        return new Connection(serverAddress, config, executor);
-    }
-
-
-    @Override
-    public PooledObject<Connection> wrap(final Connection con) {
-        return new DefaultPooledObject<>(con);
-    }
-
-    @Override
-    public boolean validateObject(final ServerAddress key, final PooledObject<Connection> p) {
-        ChannelFuture command = p.getObject().command(NSQCommand.instance("NOP"));
-        return command.awaitUninterruptibly().isSuccess();
-    }
-
-    @Override
-    public void destroyObject(final ServerAddress key, final PooledObject<Connection> p) throws Exception {
-        p.getObject().close();
-    }
-}
